# configs/swagger.yaml
openapi: 3.0.3
info:
  title: API Sgod Web
  contact:
    email: sgod.ai@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 2.0.0
servers:
  - url: http://localhost:4095/api/
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: refresh token
      description: Refresh Token
      operationId: Refresh Token
      responses:
        '200':
          description: Refresh Token successfully.
        '401':
          description: Unauthorized
      security:
        - Bearer: string
...
  /categories:
    get:
      tags:
        - Home
      summary: read categories
      description: read info Homepage.
      parameters:
        - name: sort
          in: query
          description: ID of Share to return
          schema:
            type: string
      responses:
        '200':
          description: Get all categories success
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
...
  /admin/categories:
    get:
      tags:
        - Categories
      summary: read categories
      description: read categories.
      parameters:
        - name: sort
          in: query
          description: ID of Share to return
          schema:
            type: string
      responses:
        '200':
          description: Get all categories success
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: string
    post:
      tags:
        - Categories
      summary: Create new categories
      description: This can only be done by the logged in user.
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Categories'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categories'
            application/xml:
              schema:
                $ref: '#/components/schemas/Categories'
      security:
        - Bearer: string
...
components:
  schemas:
    Categories:
      type: object
      properties:
        name:
          type: string
          default: IOT
        parent:
          type: array
          default: ['633f85e83a5cac6e4463c804']
...
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header